Fix up index, parser, and test files
