+++ index.js
@@ -2,200 +2,5 @@
 	MIT License http://www.opensource.org/licenses/mit-license.php
 	Author Tobias Koppers @sokra
 */
-var path = require('path')
-var dirname = path.dirname
-var loaderUtils = require('loader-utils')
 
-module.exports = function (source) {
-  this.cacheable && this.cacheable()
-  var jade = require('jade')
-
-  var utils = require('jade/lib/utils')
-  var nodes = require('jade/lib/nodes')
-  var filters = require('jade/lib/filters')
-
-  var req = loaderUtils.getRemainingRequest(this).replace(/^!/, '')
-
-  var query = loaderUtils.parseQuery(this.query)
-
-  var loadModule = this.loadModule
-  var resolve = this.resolve
-  var loaderContext = this
-  var callback
-
-  var fileContents = {}
-  var filePaths = {}
-  function MyParser (str, filename, options) {
-    this._mustBeInlined = false
-    jade.Parser.call(this, str, filename, options)
-  }
-  MyParser.prototype = Object.create(jade.Parser.prototype)
-  MyParser.prototype.constructor = MyParser
-
-  var missingFileMode = false
-  function getFileContent (context, request) {
-    request = loaderUtils.urlToRequest(request, query.root)
-    var baseRequest = request
-    var filePath = filePaths[context + ' ' + request]
-    if (filePath) {
-      return filePath
-    }
-    var isSync = true
-    resolve(context, request + '.jade', function (err, _request) {
-      if (err) {
-        resolve(context, request, function (err2, _request) {
-          if (err2) {
-            return callback(err2)
-          }
-
-          request = _request
-          next()
-        })
-        return
-      }
-
-      request = _request
-      next()
-      function next () {
-        loadModule('-!' + path.join(__dirname, 'stringify.loader.js') + '!' + request, function (err, source) {
-          if (err) {
-            return callback(err)
-          }
-
-          filePaths[context + ' ' + baseRequest] = request
-          fileContents[request] = JSON.parse(source)
-
-          if (!isSync) {
-            run()
-          } else {
-            isSync = false
-          }
-        })
-      }
-    })
-    if (isSync) {
-      isSync = false
-      missingFileMode = true
-      throw new Error('continue')
-    } else {
-      return filePaths[context + ' ' + baseRequest]
-    }
-  }
-
-  MyParser.prototype.parseMixin = function () {
-    this._mustBeInlined = true
-    return jade.Parser.prototype.parseMixin.call(this)
-  }
-
-  MyParser.prototype.parseBlock = function () {
-    this._mustBeInlined = true
-    return jade.Parser.prototype.parseBlock.call(this)
-  }
-
-  MyParser.prototype.parseCall = function () {
-    this._mustBeInlined = true
-    return jade.Parser.prototype.parseCall.call(this)
-  }
-
-  MyParser.prototype.parseExtends = function () {
-    if (!callback) {
-      callback = loaderContext.async()
-    }
-    if (!callback) {
-      return jade.Parser.prototype.parseExtends.call(this)
-    }
-
-    var request = this.expect('extends').val.trim()
-    var context = dirname(this.filename.split('!').pop())
-
-    var path = getFileContent(context, request)
-    var str = fileContents[path]
-    var parser = new this.constructor(str, path, this.options)
-
-    parser.blocks = this.blocks
-    parser.contexts = this.contexts
-    this.extending = parser
-
-    return new nodes.Literal('')
-  }
-
-  MyParser.prototype.parseInclude = function () {
-    if (!callback) {
-      callback = loaderContext.async()
-    }
-    if (!callback) {
-      return jade.Parser.prototype.parseInclude.call(this)
-    }
-
-    var tok = this.expect('include')
-
-    var request = tok.val.trim()
-    var context = dirname(this.filename.split('!').pop())
-    var path = getFileContent(context, request)
-    var str = fileContents[path]
-
-    // has-filter
-    if (tok.filter) {
-      var hasFilterStr = str.replace(/\r/g, '')
-      var options = {filename: path}
-      var constantinople
-      if (tok.attrs) {
-        tok.attrs.attrs.forEach(function (attribute) {
-          options[attribute.name] = constantinople.toConstant(attribute.val)
-        })
-      }
-      hasFilterStr = filters(tok.filter, hasFilterStr, options)
-      return new nodes.Literal(hasFilterStr)
-    }
-
-    // non-jade
-    if (path.substr(-5) !== '.jade') {
-      var nonJadeStr = str.replace(/\r/g, '')
-      return new nodes.Literal(nonJadeStr)
-    }
-
-    var parser = new this.constructor(str, path, this.options)
-    parser.dependencies = this.dependencies
-
-    parser.blocks = utils.merge({}, this.blocks)
-    parser.included = true
-
-    parser.mixins = this.mixins
-
-    this.context(parser)
-    var ast = parser.parse()
-    this.context()
-    ast.filename = path
-
-    if (parser._mustBeInlined) {
-      this._mustBeInlined = true
-    }
-
-    return ast
-  }
-
-  run()
-  function run () {
-    try {
-      var tmplFunc = jade.compile(source, {
-        parser: loadModule ? MyParser : undefined,
-        filename: req,
-        self: query.self,
-        globals: ['require'].concat(query.globals || []),
-        pretty: query.pretty,
-        locals: query.locals,
-        doctype: query.doctype || 'html',
-        compileDebug: loaderContext.debug || false
-      })
-    } catch (e) {
-      if (missingFileMode) {
-        // Ignore, it'll continue after async action
-        missingFileMode = false
-        return
-      }
-      throw e
-    }
-
-    loaderContext.callback(null, JSON.stringify(tmplFunc(query.locals)))
-  }
-}
+module.exports = require('./lib/index').default
